Adjust edge cuts to give fingers more clearance + duplicate PCB for OSH Park
The tolerances of the edge cuts are somewhat large (+/- 5 mil) so nudge the edge cuts up by 5 mil. If needed, it's easy to trim the flex PCB if the fingers are too long.

Also, the PCB is too small for OSH Park so duplicate it to meet their minimum size.